12/20

+++ src/app/Http/Controllers/LikeController.php
@@ -2,33 +2,35 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Http\Request;
 use App\Models\Like;
-use App\Models\Shop;
 use Illuminate\Support\Facades\Auth;
 
 class LikeController extends Controller
 {
-    public function store($shop_id) {
+    public function store(Request $request) {
         $userId = Auth::id();
-        $shop = Shop::findOrFail($shop_id);
+        $shopId = $request->input('shop_id');
 
         $like = Like::create([
             'user_id' => $userId,
-            'shop_id' => $shop->id,
+            'shop_id' => $shopId,
         ]);
 
         return back();
     }
 
-    public function destroy($shop_id){
+    public function destroy(Request $request){
         $userId = Auth::id();
-        $shop = Shop::findOrFail($shop_id);
+        $shopId = $request->input('shop_id');
 
         $likeRecord = Like::where('user_id', $userId)
-                          ->where('shop_id', $shop->id)
-                          ->first();
+                    ->where('shop_id', $shopId)
+                    ->first();
 
-        $likeRecord->delete();
+        if ($likeRecord) {
+            $likeRecord->delete();
+        }
 
         return back();
     }
